[#16][refactoring] 테이블명 변경
