+++ src/main/java/f4/auth/domain/user/persist/entity/User.java
@@ -23,7 +23,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Entity
-@Table(name = "users")
+@Table(name = "user")
 public class User {
 
   @Id
